- Add createdOn - Add updatedOn - Override saveChanges to fill createdOn and updatedOn - Get movies from dbContext - Add Genre to movies - Seed genres - Add release date and number in stock to movie - Remove discount rate from customer table in index page - Implement movies details page [INCOMPLETE] - Display genres in movies index page

+++ vidly/Models/Customer.cs
@@ -10,7 +10,7 @@ namespace vidly.Models
     {
         public bool IsSubscribedToNewsLetter { get; set; }
         public MembershipType MembershipType { get; set; }  // called navigation property
-        public byte MembershipTypeId { get; set; }  // recognized by entityframework as foreign key
+        public int MembershipTypeId { get; set; }  // recognized by entityframework as foreign key
         
         public DateTime? Birthdate { get; set; }
     }
